examples/gcp_functions/image_pipeline/clean/requirements.txt
pillow
google-cloud-tasks
google-cloud-storage